refactor(worker): extract agent-turn failure settlement
Move the runAgentTurn catch body into settleTurnFailure so recovery,
capacity waiters, and terminal failure stay one ordered function.
